Use appui-react ToolbarItem type in ToolbarPopup component.
